Final update for 7.4.1 Beta 1 users
This is the FINAL update for the the Beta line of Drauger OS over 'update-os'.
This update moves Drauger OS users on 7.4.1 Beta 1 to 7.4.1 Beta 2.
There will be no further updates from this channel for those users, only through 'apt'.
Drauger OS 7.3.7 users will get a similar update on the 7.4.1 Stable release in a few months.
That update will be the last update through 'update-os' EVER.